Add omxplayer
Fails to compile and manual intervention is required:
Replace the Makefile and the Makefile.includes files in the omxplayer
build folder with the ones in the fixes folder.
Some more changes to those to files may be required if compile still
fails, such as making sure the path to the dbus folder is correct or
that the strip command cannot be found. The error message help as to
which part of the file needs to be changed.

omxplayer works as expected when compile succeeds. Fonts are not in the
default folder though, and are not picked up automatically by omxplayer.